+++ R/distinct_se.R
@@ -18,6 +18,9 @@
 #' @export
 #'
 distinct_se <- function(.data, groupingVars, .keep_all = FALSE) {
+  if(!(is.data.frame(.data) || dplyr::is.tbl(.data))) {
+    stop("seplyr::distinct_se first argument must be a data.frame or tbl")
+  }
   # convert char vector into spliceable vector
   groupingSyms <- rlang::syms(groupingVars)
   dplyr::distinct(.data = .data, !!!groupingSyms, .keep_all = .keep_all)