+++ R/group_by_se.R
@@ -22,6 +22,9 @@
 #' @export
 #'
 group_by_se <- function(.data, groupingVars, add = FALSE) {
+  if(!(is.data.frame(.data) || dplyr::is.tbl(.data))) {
+    stop("seplyr::group_by_se first argument must be a data.frame or tbl")
+  }
   # convert char vector into spliceable vector
   groupingSyms <- rlang::syms(groupingVars)
   dplyr::group_by(.data = .data, !!!groupingSyms, add = add)