+++ R/group_indices_se.R
@@ -17,6 +17,9 @@
 #' @export
 #'
 group_indices_se <- function(.data, groupingVars, add = FALSE) {
+  if(!(is.data.frame(.data) || dplyr::is.tbl(.data))) {
+    stop("seplyr::group_indices_se first argument must be a data.frame or tbl")
+  }
   # convert char vector into spliceable vector
   groupingSyms <- rlang::syms(groupingVars)
   dplyr::group_indices(.data = .data, !!!groupingSyms, add = add)