+++ R/summarize_nse.R
@@ -27,6 +27,9 @@
 #' @export
 #'
 summarize_nse <- function(.data, ..., env = parent.frame()) {
+  if(!(is.data.frame(.data) || dplyr::is.tbl(.data))) {
+    stop("seplyr::summarize_nse first argument must be a data.frame or tbl")
+  }
   # convert char vector into spliceable vector
   # from: https://github.com/tidyverse/rlang/issues/116
   summarizeTerms <- substitute(list(...))