+++ R/summarize_se.R
@@ -23,6 +23,9 @@
 #' @export
 #'
 summarize_se <- function(.data, summarizeTerms,  env=parent.frame()) {
+  if(!(is.data.frame(.data) || dplyr::is.tbl(.data))) {
+    stop("seplyr::summarize_se first argument must be a data.frame or tbl")
+  }
   # convert char vector into spliceable vector
   # from: https://github.com/tidyverse/rlang/issues/116
   summarizeQ <- lapply(summarizeTerms,